feat: :sparkles: Added pagination utilities
It is now possible to paginate the result of the transactions and rules getter queries. When one between "limit" and "page" parameter is set, the pagination is applied (both parameter must be set, otherwise it will throw an exception). If none of these parameters are set, the query will return all the entries. If the page is greater than the maximum number of pages (calculated using the limit parameter), it will throw an exception specifying the range of possible pages available.

No breaking changes since the new parameters are optional